+++ packages/ui/src/views/secrets/index.ts
@@ -1,1 +1,2 @@
 export * from './new-secret/new-secret-form'
+export * from './secrets-header'
